# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DHP Enterprise Service API",
    "description": "Swagger definition for Digital Health Pass Enterprise Service API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/v1/enterprise",
      "description": "VPC"
    },
    {
      "url": "/",
      "description": "Sandbox"
    }
  ],
  "paths": {
...
    "/organization/{orgId}": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Get organization",
        "description": "Get details of specified organization",
        "operationId": "getOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Organization ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "organizations"
        ],
        "summary": "Delete organization",
        "description": "Delete specified organization",
        "operationId": "deleteOrganization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Organization ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/credentials/verify": {
      "post": {
        "tags": [
          "credentials"
        ],
        "summary": "Verify credentials",
        "description": "Verify credentials for holder by email",
        "operationId": "verifyCredentials",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Details of credentials to verify including organization ID and holder email",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialHolder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully verified credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidGetRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Organization": {
        "type": "object",
        "required": [
          "orgId",
          "config"
        ],
        "properties": {
          "orgId": {
            "type": "string",
            "description": "Organization ID"
          },
          "config": {
            "type": "object",
            "description": "Organization configuration properties"
          }
        }
      },
      "OrganizationResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Organization"
              },
              {
                "type": "object",
                "properties": {
                  "createdAt": {
                    "type": "string"
                  },
                  "updatedAt": {
                    "type": "string"
                  }
                }
              }
            ]
          }
        }
      },
...
      "CredentialHolder": {
        "type": "object",
        "required": [
          "orgId",
          "email"
        ],
        "properties": {
          "orgId": {
            "type": "string",
            "description": "Organization ID"
          },
          "email": {
            "type": "string",
            "description": "Holder email"
          }
        }
      },
...
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "BasicMessage": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
...
      "InvalidGetRequestError": {
        "description": "Invalid fields in request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Logged in user is unauthorized to make this request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      }
    }
  }
}